feat: add dashboard landing page
- Moved `automatic/page.tsx` to support flow control with solenoid valves and setpoint management.
- Integrated Grafana iframe for visualizing flow rate data
- Updated navigation to include new "Dashboard" and "Manual Control" routes for improved user access.
- Modified icons in `icons.css` to include the "category" icon for the dashboard.

This prepares for the classification page as dashboard

issue #40
