+++ app/ui/icons.css
@@ -1,1 +1,1 @@
-@import url(https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&icon_names=autoplay,gas_meter,play_arrow,settings_slow_motion,valve,water_drop,water_pump);
+@import url(https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&icon_names=autoplay,category,gas_meter,play_arrow,settings_slow_motion,valve,water_drop,water_pump);
